Add storage module
Store items in local storage.

The items collections is stored in the `items.all` namespace, under the
form of an object of shape { [id: string]: (title: string), ... }

The individual items are stored in their own namespace of the form
`items.${id}` and of value `content: string`, content being the
encrypted output of the item's real content.

Following methods are exported:
- getAllItems()
- getItem(id)
- deleteItem(id)
- saveItem(id, title, content)